Issue SB-219 feat:changes for local actor impl
