+++ actors/src/main/java/org/sunbird/learner/actors/BackgroundRequestRouterActor.java
@@ -13,7 +13,9 @@ import org.sunbird.common.request.Request;
 import org.sunbird.common.responsecode.ResponseCode;
 import org.sunbird.learner.actors.bulkupload.BulkUploadBackGroundJobActor;
 import org.sunbird.learner.audit.impl.ActorAuditLogServiceImpl;
+import org.sunbird.metrics.actors.CourseMetricsBackgroundActor;
 import org.sunbird.metrics.actors.MetricsBackGroundJobActor;
+import org.sunbird.metrics.actors.OrganisationMetricsBackgroundActor;
 
 public class BackgroundRequestRouterActor extends UntypedAbstractActor {
 
@@ -25,6 +27,10 @@ public class BackgroundRequestRouterActor extends UntypedAbstractActor {
   private ActorRef metricsBackGroungJobActor;
   
   private ActorRef auditLogManagementActor;
+  
+  private ActorRef organisationMetricsBackgroundActor;
+  
+  private ActorRef courseMetricsBackgroundActor;
  
 
 
@@ -34,6 +40,8 @@ public class BackgroundRequestRouterActor extends UntypedAbstractActor {
   private static final String BULK_UPLOAD_BACKGROUND_ACTOR = "bulkUploadBackGroundJobActor";
   private static final String METRICS_BACKGROUND_ACTOR = "metricsBackGroungJobActor";
   private static final String AUDIT_LOG_MGMT_ACTOR = "auditLogManagementActor";
+  private static final String ORG_METRICS_BACKGROUND_ACTOR = "organisationMetricsBackgroundActor";
+  private static final String COURSE_METRICS_BACKGROUND_ACTOR = "courseMetricsBackgroundActor";
 
 
   /**
@@ -55,6 +63,14 @@ public class BackgroundRequestRouterActor extends UntypedAbstractActor {
     auditLogManagementActor = getContext().actorOf(
         FromConfig.getInstance().props(Props.create(ActorAuditLogServiceImpl.class)),
         AUDIT_LOG_MGMT_ACTOR);
+    
+    organisationMetricsBackgroundActor = getContext().actorOf(
+        FromConfig.getInstance().props(Props.create(OrganisationMetricsBackgroundActor.class)),
+        ORG_METRICS_BACKGROUND_ACTOR);
+    
+    courseMetricsBackgroundActor = getContext().actorOf(
+        FromConfig.getInstance().props(Props.create(CourseMetricsBackgroundActor.class)),
+        COURSE_METRICS_BACKGROUND_ACTOR);
 
     initializeRouterMap();
   }
@@ -68,8 +84,8 @@ public class BackgroundRequestRouterActor extends UntypedAbstractActor {
     routerMap.put(ActorOperations.FILE_GENERATION_AND_UPLOAD.getValue(), metricsBackGroungJobActor);
     routerMap.put(ActorOperations.UPDATE_USER_INFO_ELASTIC.getValue(), backgroundJobManager);
     routerMap.put(ActorOperations.UPDATE_USER_ROLES_ES.getValue(), backgroundJobManager);
-    routerMap.put(ActorOperations.PROCESS_DATA.getValue(), backgroundJobManager);
-    routerMap.put(ActorOperations.FILE_GENERATION_AND_UPLOAD.getValue(), backgroundJobManager);
+    routerMap.put(ActorOperations.PROCESS_DATA.getValue(), metricsBackGroungJobActor);
+    routerMap.put(ActorOperations.FILE_GENERATION_AND_UPLOAD.getValue(), metricsBackGroungJobActor);
     routerMap.put(ActorOperations.ADD_USER_BADGE_BKG.getValue(), backgroundJobManager);
     routerMap.put(ActorOperations.UPDATE_USR_COURSES_INFO_ELASTIC.getValue(), backgroundJobManager);
     routerMap.put(ActorOperations.UPDATE_USR_COURSES_INFO_ELASTIC.getValue(), backgroundJobManager);
@@ -84,6 +100,9 @@ public class BackgroundRequestRouterActor extends UntypedAbstractActor {
     routerMap.put(ActorOperations.INSERT_COURSE_BATCH_ES.getValue(), backgroundJobManager);
     routerMap.put(ActorOperations.SEARCH_AUDIT_LOG.getValue(), auditLogManagementActor);
     routerMap.put(ActorOperations.PROCESS_AUDIT_LOG.getValue(), auditLogManagementActor);
+    routerMap.put(ActorOperations.ORG_CREATION_METRICS_DATA.getValue(), organisationMetricsBackgroundActor);
+    routerMap.put(ActorOperations.ORG_CONSUMPTION_METRICS_DATA.getValue(), organisationMetricsBackgroundActor);
+    routerMap.put(ActorOperations.COURSE_PROGRESS_METRICS_DATA.getValue(), courseMetricsBackgroundActor);
   }
