+++ actors/src/main/java/org/sunbird/metrics/actors/BaseMetricsActor.java
@@ -26,7 +26,6 @@ import org.sunbird.common.exception.ProjectCommonException;
 import org.sunbird.common.models.response.Response;
 import org.sunbird.common.models.util.HttpUtil;
 import org.sunbird.common.models.util.JsonKey;
-import org.sunbird.common.models.util.LoggerEnum;
 import org.sunbird.common.models.util.ProjectLogger;
 import org.sunbird.common.models.util.ProjectUtil;
 import org.sunbird.common.models.util.PropertiesCache;
@@ -42,7 +41,6 @@ public abstract class BaseMetricsActor extends UntypedAbstractActor {
   
   private static ObjectMapper mapper = new ObjectMapper();
 
-  protected abstract Map<String, Object> getViewData(String id, Object data);
   public static final String startDate = "startDate";
   public static final String endDate = "endDate";
   public static final String startTimeMilis = "startTimeMilis";
