+++ actors/src/main/resources/application.conf
@@ -346,6 +346,16 @@ BackGroundLocalMWConfig {
             router = round-robin-pool
             nr-of-instances = 5
           }
+        /BackgroundRequestRouterActor/organisationMetricsBackgroundActor
+          {
+            router = round-robin-pool
+            nr-of-instances = 5
+          }
+        /BackgroundRequestRouterActor/courseMetricsBackgroundActor
+          {
+            router = round-robin-pool
+            nr-of-instances = 5
+          }
       }
      
     }
@@ -382,6 +392,16 @@ BackgroundRemoteMWConfig {
             router = round-robin-pool
             nr-of-instances = 5
           }
+        /BackgroundRequestRouterActor/organisationMetricsBackgroundActor
+          {
+            router = round-robin-pool
+            nr-of-instances = 5
+          }
+        /BackgroundRequestRouterActor/courseMetricsBackgroundActor
+          {
+            router = round-robin-pool
+            nr-of-instances = 5
+          }
       }
      
     }
@@ -410,7 +430,7 @@ ActorConfig {
       enabled-transports = ["akka.remote.netty.tcp"]
       netty.tcp {
       port = 8099
-      hostname = learner-service
+      #hostname = learner-service
       message-frame-size =  30000000b
       send-buffer-size =  30000000b
       receive-buffer-size =  30000000b
